+++ qgreenland/config/helpers/layers/sea_ice_age.py
@@ -2,7 +2,9 @@ import json
 from typing import Literal
 
 from qgreenland.config.datasets.seaice import seaice_age as dataset
-from qgreenland.config.helpers.steps.build_overviews import build_overviews
+from qgreenland.config.helpers.steps.compress_and_add_overviews import (
+    compress_and_add_overviews,
+)
 from qgreenland.config.helpers.steps.gdal_edit import gdal_edit
 from qgreenland.config.helpers.steps.warp_and_cut import warp_and_cut
 from qgreenland.config.project import project
@@ -70,9 +72,10 @@ def sea_ice_age_layer(year: int, age_type: AgeType) -> Layer:
                     '-tr', '12500', '12500',
                 ],
             ),
-            *build_overviews(
+            *compress_and_add_overviews(
                 input_file='{input_dir}/warped_and_cut.tif',
                 output_file='{output_dir}/overviews.tif',
+                dtype_is_float=False,
             ),
         ],
     )