+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# NEXT_VERSION
+
+- Cast RACMO promice mask layers to `Byte` data type.
+
+
 # v2.0.0alpha3 (2021-11-23)
 
 - Improved raster compression.