+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v2.0.0alpha3 (2021-11-23)
 
 - Improved raster compression.
+- Update Bedmachine dataset metadata to reflect new v4 data.
 
 
 # v2.0.0alpha2 (2021-11-22)